refactor: separate timeline track content into new file
